fix(r/demo/boards): update the user link in displayAddressMD to correct path (#1525)
This updates r/demo/boards to fix the link when clicking in a username.
To see the problem, visit https://test3.gno.land/r/demo/boards:testboard
. Click on the username `@administrator`. The result is "404 page not
found". The problem is that the URL created by `displayAddressMD` is:

https://test3.gno.land/r/users:administrator

but it should be

https://test3.gno.land/r/demo/users:administrator
 
The first commit changes `displayAddressMD` to the correct
"r/demo/users". The second commit updates the expected test outputs.

---------

Signed-off-by: Jeff Thompson <jeff@thefirst.org>