fix(gnofaucet): fix transactions by querying account number and sequence each time (#1493)
Currently, `gnofaucet` [queries the blockchain
once](https://github.com/gnolang/gno/blob/c3a4993335f8881989eb49e4efea6a3c1310061d/gno.land/cmd/gnofaucet/serve.go#L191-L209)
for the account number and sequence number, and [increments the sequence
number](https://github.com/gnolang/gno/blob/master/gno.land/cmd/gnofaucet/serve.go#L327)
each time that it sends coins. We are using `gnofaucet` with `gnodev`.
When `gnodev` restarts the chain, it also resets the sequence number.
This means that we also have to restart `gnofaucet` (to re-query the
sequence number), which is inconvenient for the development workflow.

If you agree that it would be nice for `gnofaucet` to continue to work
after `gnodev` does a reset, then this pull request moves the query for
the account number and sequence number into the `sendAmountTo` function
so that it queries each time it signs a transaction. Now `gnofaucet`
continues to function even if the sequence number is unexpectedly
changed (because the blockchain is reset or because some other
transaction is signed with the same key).

Signed-off-by: Jeff Thompson <jeff@thefirst.org>