docs: s/Gnolang/Gno, s/Golang/Go (#1548)
They're Gno and Go. The "lang" part should be used only for
disambiguation. Same as we don't call C "Clang", Java "Javalang" and so
on.

Arguably, we could drop "gnolang" completely outside maybe of internal
usage in the code, as contrary to Go, "Gno" is not a common verb in the
english language.

---------

Co-authored-by: Leon Hudak <33522493+leohhhn@users.noreply.github.com>