chore(CODEOWNERS): prefix directories with root slash (#1573)
Pull request #1572 marks the codeowner as @moul, in spite of what the
configuration might suggest to a human reader.

[Reading through GitHub's
docs](https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-code-owners#example-of-a-codeowners-file),
this seems to come from the fact that if directories are not prefixed
with a `/`, they are considered as applying to all files or directories
with that name. (ie. `foo` matches `/foo` but also `/src/foo`).

I've added `/` to all directories in CODEOWNERS to fix this.