fix(tm2): load config defaults if not specified in existing config (#1544)
## Description

This PR fixes the flow of loading Gno configuration files. 
Previously, configuration files were loaded without applying default
values, causing to funky behavior.

Additionally, this PR drops the config TOML template, instead relying on
the TOML package itself to marshal.

Issue spotted in #1238 

Note:
I've tried to keep the original functionality of
`LoadOrMakeConfigWithOptions`, because if this is changed or broken up
in its current state, things start breaking across dimensions.

Funky behavior that's fixed:
The `EventStore` fields were not being saved in the original config
(since we utilized a template for it, and this was not present in the
template). Now, config files should properly save on disk the event
store configuration.

<details><summary>Contributors' checklist...</summary>

- [x] Added new tests, or not needed, or not feasible
- [x] Provided an example (e.g. screenshot) to aid review or the PR is
self-explanatory
- [x] Updated the official documentation or not needed
- [x] No breaking changes were made, or a `BREAKING CHANGE: xxx` message
was included in the description
- [x] Added references to related issues and PRs
- [ ] Provided any useful hints for running manual tests
- [ ] Added new benchmarks to [generated
graphs](https://gnoland.github.io/benchmarks), if any. More info
[here](https://github.com/gnolang/gno/blob/master/.benchmarks/README.md).
</details>