docs: add edit page link and info (#1629)
Related to #1505,This PR aims to add the edit page option in order to
link the source `md` file to docs page. So readers will be invited to
edit the page to improve it. It also add the last update date, so
readers can know if the page is up to date.