fix(gnovm): don't treat `[]byte("xxx")` as a constant expression (#1597)
this is a fix to #1570 .

The underlying issue in #1570 stems from how []byte("xxx") is handled:
it's treated as a constant, yet its value can still be altered. This
mutability leads to unintended persistence of its state across loop
iterations.

The proposed resolution involves modifying the preprocessing stage to
avoid converting an untyped string directly into a []byte type.