Implement a go expression parser for the 'print' command.
debugEvalExpr evaluates a Go expression in the context of the VM and returns
the corresponding value, or an error.
The supported expression syntax is a small subset of Go expressions:
basic literals, identifiers, selectors, index expressions, or a combination
of those are supported, but none of function calls, arithmetic, logic or
assign operations, type assertions of convertions.