fix:  unify the precompile output to make it easier to parse and read.  (#1670)
<!-- please provide a detailed description of the changes made in this
pull request. -->
Fix #1636 

Under the hood, use `scanner.ErrorList` from the stdlib all the way to
store all errors with the filename and the position in the file. Note
that kind of error is already returned by `parser.ParserFile`.

For the `go build` output, instead of printing it like it is, parse it
and shift it to represent the related gno file.

Run the tests:
```
$ go test ./gnovm/cmd/gno -v -run Test_ScriptsPrecompile
```

<details><summary>Contributors' checklist...</summary>

- [x] Added new tests, or not needed, or not feasible
- [ ] Provided an example (e.g. screenshot) to aid review or the PR is
self-explanatory
- [ ] Updated the official documentation or not needed
- [ ] No breaking changes were made, or a `BREAKING CHANGE: xxx` message
was included in the description
- [x] Added references to related issues and PRs
- [x] Provided any useful hints for running manual tests
- [ ] Added new benchmarks to [generated
graphs](https://gnoland.github.io/benchmarks), if any. More info
[here](https://github.com/gnolang/gno/blob/master/.benchmarks/README.md).
</details>