feat(examples): add package cford32, add method `seqid.ID.String` (#1572)
This PR adds a new package to examples, `cford32`, meant primarily to be
used in package `seqid` as an AVL- and human-friendly ID, which
implements an encoding scheme with the base32 encoding scheme [specified
by Douglas Crockford](https://www.crockford.com/base32.html). It
additionally implements a `uint64` encoding scheme I created, which
encodes "tiny" (< 17 billion) values as 7-byte strings, and can encode
the full `uint64` range with 13 bytes.

The package is largely a fork of Go's `encoding/base32`, intentionally
forked to have a very familiar API, while needing to be forked to
implement some distinctive features of the encoding (like case
insensitivity, and mapping in decoding all of the symbols `l L i I 1` to
the same value).

The necessity of this package comes from a solution that I implemented
in GnoChess:


https://github.com/gnolang/gnochess/blob/9aa813fbb86fec377a85fc4528411d652fc780ff/realm/chess.gno#L286-L295

Essentially, GnoChess used simple sequential IDs for its saved entities
(like games). To work well with AVL's sorted keys, it padded the
generated strings to the left with up to 9 zeroes. This, of course,
breaks for values `>= 1e10` (10 billion), as `("2" + "000000000") >
("10" + "000000000")`.