refactor(gnovm): rename precompiler to transpiler, move to own package (#1681)
Part 1 of several _transpiler_ refactors I intend to perform.

This PR tackles one of the first pet peeves I have about the
`precompile` tool, which is that the name is not intuitive to what it
does.

Transpile is a name that not everyone likes. Ultimately, converting one
source code to another is also an act of "compilation"; in fact, the
TypeScript tool that converts it into JavaScript is called the
"TypeScript Compiler". The distinction between transpiler and compiler
is blurry; but I think it is useful to call what we have as a
_transpiler_ to outline that it is doing a "source-to-source"
translation, which I don't think a name like `compile` or `precompile`
could.

Additionally, this PR moves the source files for the transpiler into its
own package, gnovm/pkg/transpiler.

One question is to whether the "constants" of the precompiler, like
`GnoRealmPkgsPrefixBefore`, should deserve their own package or not.