refactor(tm2): make rpc client not depend on goleveldb (#1603)
![image](https://github.com/gnolang/gno/assets/4681308/761f1109-3e0d-4846-b331-3c60b6e803c0)


![image](https://github.com/gnolang/gno/assets/4681308/312b9226-49fa-48ac-b280-f4d704a341df)

The PR depends on #1602 but it's good for review. For ease of reviewing
I put that PR's branch as the base branch; once it is merged I will
rebase on master.

A bunch of changes to get there:

- Split `proxy` into `appconn` and `proxy`. `proxy` has a dependency on
`kvstore`, which then depends on pkg/db and goleveldb.
- Tools for inspecting dependency chains:
- `go list -f '{{ join .Deps "\n" }}'
github.com/gnolang/gno/tm2/your/pkg`
  - [`depth`](https://github.com/KyleBanks/depth)

The original point of the PR was to not include an entire key-store
database as part of `cmd/gno`, but really the better side-effect is that
the dependency is no longer in the rpc package. Sadly, `gnoclient` still
imports `goleveldb`, but I'll leave understanding why that is the case
for another refactor-investigation :)