docs: add new Prerequisites (#1722)
<!-- please provide a detailed description of the changes made in this
pull request. -->
I propose this change in the prerequisites of the
[getting-started](https://docs.gno.land/getting-started/local-setup)
because @kazai777 and I (mac apple silicon users) encountered a `command
not found` with the given commands. And by checking with `echo $GOPATH`,
the path was empty. We noticed that in the newer versions of Go, one
must manually add the path.

However, thanks to @mous1985 , we saw that on Ubuntu, it was set
automatically.

Warning -> this also needs to be verified on Windows.
<details><summary>Contributors' checklist...</summary>

- [x] Added new tests, or not needed, or not feasible
- [x] Provided an example (e.g. screenshot) to aid review or the PR is
self-explanatory
- [x] Updated the official documentation or not needed
- [x] No breaking changes were made, or a `BREAKING CHANGE: xxx` message
was included in the description
- [x] Added references to related issues and PRs
- [x] Provided any useful hints for running manual tests
- [ ] Added new benchmarks to [generated
graphs](https://gnoland.github.io/benchmarks), if any. More info
[here](https://github.com/gnolang/gno/blob/master/.benchmarks/README.md).
</details>

---------

Co-authored-by: théo dub <theodub@MacThox.local>
Co-authored-by: Leon Hudak <33522493+leohhhn@users.noreply.github.com>