feat: allow disabling consensus WAL (#1723)
<!-- please provide a detailed description of the changes made in this
pull request. -->
This allows for the consensus state's WAL to be disabled. This is
desirable for testing against a running node when we don't care about
the WAL -- it avoids potential permissions issues by not writing
anything to disk. This allows for our integration testing package to be
imported and used outside of the gno repository.
<details><summary>Contributors' checklist...</summary>

- [ ] Added new tests, or not needed, or not feasible
- [ ] Provided an example (e.g. screenshot) to aid review or the PR is
self-explanatory
- [ ] Updated the official documentation or not needed
- [ ] No breaking changes were made, or a `BREAKING CHANGE: xxx` message
was included in the description
- [ ] Added references to related issues and PRs
- [ ] Provided any useful hints for running manual tests
- [ ] Added new benchmarks to [generated
graphs](https://gnoland.github.io/benchmarks), if any. More info
[here](https://github.com/gnolang/gno/blob/master/.benchmarks/README.md).
</details>

---------

Co-authored-by: Morgan <git@howl.moe>