chore(gnoweb/docs): better website connectivity (#1763)
## Description

This PR replaces the "Start" button with the "Docs" button on the
gno.land home page navbar, and adds a "Home" button to the docs navbar
to interlink the websites in a nice manner. It also fixes a small typo
in the GoR page.

<details><summary>Contributors' checklist...</summary>

- [x] Added new tests, or not needed, or not feasible
- [x] Provided an example (e.g. screenshot) to aid review or the PR is
self-explanatory
- [x] Updated the official documentation or not needed
- [x] No breaking changes were made, or a `BREAKING CHANGE: xxx` message
was included in the description
- [x] Added references to related issues and PRs
- [x] Provided any useful hints for running manual tests
- [ ] Added new benchmarks to [generated
graphs](https://gnoland.github.io/benchmarks), if any. More info
[here](https://github.com/gnolang/gno/blob/master/.benchmarks/README.md).
</details>