feat(examples): add memeland (#1751)
## Description

This PR adds the `p/demo/memeland` & `r/demo/memeland` package & realm
to the examples folder. Adding this will be the way we "Deploy" the
Memeland realm to the Portal Loop.

The UI and the most up to date code can be found
[here](https://github.com/gnolang/memeland). Waiting on full release and
will copy code over from the other repo.

The only thing currently missing are public realm functions tested. The
package is fully tested.

<details><summary>Contributors' checklist...</summary>

- [x] Added new tests, or not needed, or not feasible
- [x] Provided an example (e.g. screenshot) to aid review or the PR is
self-explanatory
- [x] Updated the official documentation or not needed
- [x] No breaking changes were made, or a `BREAKING CHANGE: xxx` message
was included in the description
- [x] Added references to related issues and PRs
- [x] Provided any useful hints for running manual tests
- [ ] Added new benchmarks to [generated
graphs](https://gnoland.github.io/benchmarks), if any. More info
[here](https://github.com/gnolang/gno/blob/master/.benchmarks/README.md).
</details>