docs: add Playground section to "Getting Started" (#1773)
## Description

This PR adds the Playground to the Getting Started section in the docs.
The idea is to put forward the Playground as the first place for people
to go and learn about Gno, instead of having to set up a local chain and
all of the things that follow.

I suggest reviewing this PR by running a local version of the docs, with
`yarn start` in the `misc/docusaurus` folder. (when deployment preview
@albttx)

This is one of 2 PRs that will update the Getting started section in the
Docs. The next PR will update the Local setup section to put `gnodev` as
the first option to get started.

Here is the latest version of the PR:
https://www.loom.com/share/636afb0d48ed4bf7ac8043b17463b8fd?sid=0c055f69-a5a8-4ddf-8d42-f550db17dbe3

<details><summary>Contributors' checklist...</summary>

- [x] Added new tests, or not needed, or not feasible
- [x] Provided an example (e.g. screenshot) to aid review or the PR is
self-explanatory
- [x] Updated the official documentation or not needed
- [x] No breaking changes were made, or a `BREAKING CHANGE: xxx` message
was included in the description
- [x] Added references to related issues and PRs
- [x] Provided any useful hints for running manual tests
- [ ] Added new benchmarks to [generated
graphs](https://gnoland.github.io/benchmarks), if any. More info
[here](https://github.com/gnolang/gno/blob/master/.benchmarks/README.md).
</details>

---------

Co-authored-by: Danny <salman.danny03@hotmail.com>